refactor: Extract version management into VersionManagementService module
- Created src/tools/modules/services/version-management-service.ts (155 lines)
- Extracted version checking and update methods:
  - checkForUpdates (main orchestrator)
  - getCurrentVersion
  - getLatestVersionInfo
  - compareVersions
  - shouldCheckForUpdates
- All methods now static in VersionManagementService class
- Updated all calls to use VersionManagementService
- Fixed internal reference to use VersionManagementService.compareVersions
- Removed old private method implementations (156 lines)

Impact:
- tools/index.ts reduced from 1,147 to 991 lines (156 lines removed)
- All 296 tests passing ✅
- No TypeScript errors ✅

🎯 PHASE 2 TARGET ACHIEVED: File is now 991 lines (< 1,000 target)
Overall: Reduced from original 5,169 to 991 lines (80.8% reduction)
